More Numbers sections 3 and 4

+++ app/src/main/java/org/onebillion/onecourse/mainui/OC_SectionController.java
@@ -505,7 +505,7 @@ public class OC_SectionController extends OBSectionController {
         return p;
     }
 
-    public void moveScenePointer(PointF point, int angle, float time, String audio, int index, float wait) throws Exception
+    public void moveScenePointer(PointF point, float angle, float time, String audio, int index, float wait) throws Exception
     {
         movePointerToPoint(point,angle,time,true);
         playAudioScene(currentEvent(),audio,index);
